# specs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Sinaps MVP API
  version: v1.0.0
  description: API for Sinaps real-time communication platform
paths:
  /auth/apple:
    post:
      summary: Verify Apple ID token and issue JWT
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                token:
                  type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  jwt:
                    type: string
                  livekitToken:
                    type: string
  /auth/login:
    post:
      summary: Login with credentials
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                password:
                  type: string
      responses:
        '200':
          description: OK
...